LU-12597 tests: return comma-separated mdts_nodes()
Start to return comma-separated MDTs list from mdts_nodes(),
to avoid the redundant calling of comma_list() for each user.

Fix the few places that do not call comma_list() on mdts_nodes()
output afterward, or treat the output as a space-separated list.

Later patches will clean up all of the callers of mdts_nodes,
but there are too many places it is used to do in a single patch.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iddd3278321e7552dc2afd5fbb9f48033af3ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57194
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>