LU-18501 tests: sanity.sh test_56oe fix time range
test_56oe was likely failing due to the test running slow, resulting
in all of the files escaping the time range before they are found
with 'lfs find -ctime -Tupper -ctime +Tlower' so the bounds are now
calculated after all files have been created meaning that even if
there are large gaps between the creation of each file they will
still be within the new bounds.

Just in case the test is really extra slow, this test will be skipped
to avoid having to retest for the 0.01% of the time that it happens.

Test-Parameters: trivial testlist=sanity env=ONLY=56oe,ONLY_REPEAT=100
Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: If3f737146f1a1d110873c5dd3a6ece60604e54ac
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57201
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>