LU-15420 llite: use op_data to store context states
The function ll_lookup_it() uses pointers to both secctx
and encctx states since that data can either come from the
server or is created locally on the client. Here we move
the code to use the pointer already in the struct md_op_data.
The reason for this change is that the fscrypt API needs
this information when setting a encrypted context.

Change-Id: Ia46ef093beb08a960fda6493326c92fd905b02f9
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56353
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Timothy Day <timday@amazon.com>