LU-13118 llite: add kobject to struct config_llog_instance
For the case of PARAM_LLITE in class_process_config() special
handling was implemented to locate the super block which
containes the struct kobject needed to send an uevent. This
special handling depends on naming containing an internal
kernel address which is also used to build the sysfs tree.
To fix this ASLR violation we need to move away from using
this internal address but it will break our special
handling. The solution is to add kobject of llite to the
struct config_llog_instance that can be used by the
class_process_config() so the special handling can go away.
Plus the change allows the removal of lsi_kobj.

Change-Id: Idf1e5a87e9d7b1b6448877d8bc21baf7f823dfeb
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56850
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>