LU-8066 obdecho: remove last user of lprocfs_uuid_seq_show
The function lprocfs_obd_setup() will by default create a set
of sysfs files. One of those files is uuid. With this being
the case obdecho is creating two uuid files, one in sysfs and
the other in debugfs. We don't need the debugfs version so
remove it. With this change we can remove the last uuid
debugfs / procfs version. Also procfs setup for obdecho is
no longer needed.

Change-Id: I741b8f98ebac83da95255cb6c26b70f9fbce3ef7
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57400
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>