LU-18540 lnet: Fix lnetctl ping timeout parsing
The src_nid argument to yaml_lnet_ping() is never null, so we need to
check it with nid_addr_is_set() to determine whether the source key
should be added to the ping command yaml.

sanity-lnet/256 is updated to verify that all ping commands complete
successfully. This verifies that lnetctl ping --timeout works
correctly.

Test-Parameters: trivial testlist=sanity-lnet
Fixes: 5ddc054540 ("LU-17629 utils: support hostname with lustre_lnet_parse_nid_range()")
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Id6617b38edc68e49800ae37982390e17bc95e141
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57405
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>