LU-18421 utils: Do not inherit stripe size from DOM component
When a DOM file is created with multiple components,
the size of the first component is inherited by the
other components if their sizes are not explicitly specified.

This fix ensures that the size of subsequent components is
set to the default size when no size is provided.

New test added in
sanity-dom.sh:
Added test 8 to validate the creation of a default
stripe size when no stripe size is explicitly specified.

Modified test--
sanity-pfl.sh:
modified 21b test so that stripe size is 1MB
as test assumes 1MB size while doing find

Test-Parameters: testlist=sanity-dom
HPE-bug-id: LUS-12585
Signed-off-by: Rajeev Mishra <rajeevm@hpe.com>
Change-Id: I3388c289084769a96d8b87ae3a8268e5eb19a67e
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57501
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>
Reviewed-by: Mikhail Pershin <mpershin@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>