LU-18357 ptlrpc: Add fileset ID handling on the nodemap IAM
For handling multiple filesets in the nodemap's IAM cluster idx space,
the cluster idx subids must be associated with a fileset id and vice
versa. This allows to place fileset fragments in the correct subid
range and to assign retrieved fragments to the correct fileset.

This patch adds these calculations for a single fileset with the work
for multiple alternate filesets pending.

Test-Parameters: trivial testlist=sanity-sec
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I7bef5218ebb38d57d0e3aa6aab499111733eaef7
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57527
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>