LU-18357 ptlrpc: Support add and del alternate filesets
This patch adds the API and functionality to add and delete alternate
filesets to a nodemap. This is implemented by using the previously
added fileset rb tree to store alternate filesets in the in-memory
nodemap struct. Both, the rb tree and nodemap IAM is synchronized.
Note, the primary fileset is handled outside the rb tree.

Consistency between both data structures is guaranteed as follows:
When adding a fileset, it is first inserted into the rb tree,
generating a free fileset ID which is then used to insert the fileset
to the nodemap IAM at the subid calculated from the fileset ID. On
reading the IAM index pages (on nodemap synchronization or restart),
the fileset ID is derived from the subid and used on inserting the
fileset into the rb tree. Therefore, free fileset IDs can be managed
by the rb tree.

The following commands are added/modified in this patch:

- (new) lctl nodemap_fileset_add [--alt] --nodemap .. --fileset '..'
- (new) lctl nodemap_fileset_del --nodemap .. --fileset '..'
- (modified) 'lctl get_param nodemap.<nm_name>.fileset' now prints
  both fileset types. Tests sanity-sec 27a and 27b were modified
  accordingly.
- (modified) 'lctl nodemap_set_fileset' is deprecated and now prints a
  corresponding warning on use.

This patch adds sanity-sec 27c which tests multiple fileset
persistence, varying fileset lengths, and the limit of 255 alternate
filesets.

Note, alternate filesets are not used yet when mounting a client.
Implementing the mounting rules based on both fileset types is done in
the next patch, including the testing of multiple filesets on client
mount.

Test-Parameters: trivial testlist=sanity-sec
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: Ie80428732b22892d817c937d3c60c4f477ce7f0b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56668
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>