LU-18357 ptlrpc: Multiple fileset support on client mount
This patch adds new client mounting rules when primary and alternate
filesets are present on a nodemap. Whenever a nodemap is relevant for
given client node, these mounting rules decide whether the client can
mount a given subdirectory.  The mounting rules are as follows:

1. If the nodemap is inactive or no filesets are set, there are no
fileset restrictions on the mounting subdir, i.e., no mounting
restrictions.
2. If no mounting subdir is set, the primary fileset is used and the
mounting subdir is replaced by it.
3. If any fileset (primary and alternate) fileset matches the mounting
subdir exactly or as a prefix, the client is permitted to mount the
subdir.
4. If no fileset was found in 3., the mounting subdir is appended to
the primary fileset.

In all cases (for rules 2-4), the subdir must exist to mount it. If no
primary fileset is set, rules 2 and 4 are omitted, forcing an mount
subdir that is represented by a alternate fileset. Hence, if neither
cases 2 and 4 apply, mounting the subdirectory is denied.

A new sanity-sec 27d test is added to test these mounting rules in 18
subtests.

Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I7c0b1721584bc123a69e479d4cf4cfb4d4064e42
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57609
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>