LU-13335 utils: allow projid lookup by name
Add support for "lfs project", "lfs quota" and "lfs find"  that deal
with project IDs to look up project IDs by name from /etc/projid file.

This allows an administrator-friendly interface for using projects
instead of having to remember and use the numeric projid value.

In order to avoid overloading the /etc/projid(5) file with duplicate
entries for users and groups, and to allow the use of external ID
databases management (LDAP, etc.), there is also an opt-in to specify
a username or group name in place of a project name that will use the
numeric UID or GID as the PROJID:
    'lfs project -p u:quota_usr /mnt/lustre/tfile'
    'lfs project -p g:quota_grp /mnt/lustre/tfile'

Also now shows projname for lfs quota instead of projid if available.

Modified uid2name and gid2name to use a buffer use getpwuid_r and
getgrgid to avoid unsafe calls with multi-threading.

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I7b50c4814d362d7e76ad1f82669874d7a93ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/46144
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Sergey Cheremencev <scherementsev@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>