LU-18418 utils: new lctl find_param option
'lctl find_param PATTERN' will search through all parameter names and
show those that match the extended regular expression PATTERN.
It is an equivalent to
    'lctl get_param -R "*" | grep -E PATTERN | xargs lctl get_param'

For example, 'lctl get_param debug_path' returns
    debug_path=/tmp/lustre-log
however 'lctl find_param lustre | grep debug_path' returns nothing
as "lustre" is not found in the parameter name "debug_path".

This is useful when you know a specific parameter but not the full
path or when the path has changed, like how "brw_stats" is at
"osd-ldiskfs.*.brw_stats" but it used to be at "obdfilter.*.brw_stats"

LU-14590 introduced significant slow-down to lctl list_param so a
minor fix was introduced in this patch.

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I34d45356f9b71021ff1147cf01e18ebb6182e7ce
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57675
Reviewed-by: Aurelien Degremont <adegremont@nvidia.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>