LU-17778 tests: fix conf-sanity/76d issues
There was a race between remounting a client and the MGS parameter
settings being applied, so ensure the parameter setting is rechecked
if it is not correct the first time.

Also, the parameter checking for $MOUNT2 was not necessarily checking
the right instance of the parameter on the client.  Use the instance
name for the mountpoint to ensure this is checked correctly.

Test-Parameters: trivial
Fixes: fa1bff8f6f ("LU-9399 llite: register mountpoint before process llog")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ib6acdc80b880dfc90b0c10406a0f868211433f58
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58215
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Emoly Liu <emoly@whamcloud.com>
Reviewed-by: Nushafreen Palsetia <npalsetia@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>