LU-18643 tests: Do not create subdirectory on client mount
When calling zconf_mount_clients() with a FILESET, the Lustre client
is mounted at the corresponding subdirectory specified by the FILESET.

However, when the subdirectory does not exist, it automatically
creates it transparently. This may hide bugs, e.g., when a test needs
to verify that mounting against a non-existing directory is not
possible.

This patch removes the silent creation of the directory on client
mount, so that it is the caller's responsibility that the subdirectory
exists before mounting. Currently, no tests are relying on this
functionality as they already create the subdirectory themselves.
Therefore no test needs to be modified.

Test-Parameters: mdtcount=4 mdscount=2 env=ONLY="247 413" testlist=sanity
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I900c4bff79e6b5bde541eb4e852e42cde01820e3
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57807
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>