LU-18357 ptlrpc: Support clearing all filesets from a nodemap
Deleting filesets requires the name of the fileset to delete. While
all filesets can be removed at once, this is only done implicitly when
deleting a nodemap.

To improve usability, this patch adds the "--all" argument for the
existing "lctl nodemap_fileset_del" command to delete all filesets
from a nodemap.

Sanity-sec 27c was modified to test this functionality.

Test-Parameters: trivial testlist=sanity-sec
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I30bb06c9cddd404f474223745826d7a17832ea42
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57900
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>