LU-18728 tests: use urandom to really consume ZFS space
It appears newer ZFS is using data compression by default, so reading
from /dev/zero results in files not consuming the expected amount of
space.  Instead, read from /dev/urandom for ZFS to write files in
sanity and conf-sanity to ensure they fill the OSTs, or the image
to be used for target creation, as expected.

Test-Parameters: testgroup=review-zfs env=ZFS_MKFS_OPTS="compression=on"
Signed-off-by: Bruno Faccini <bfaccini@nvidia.com>
Change-Id: I7b4e95032608d8db82c75e4b6dd1ec5beb6f8d99
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58115
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>