LU-18357 ptlrpc: Add read-only flag for nodemap filesets
When adding a fileset to a nodemap, each fileset was previously set to
read-write except if the whole nodemap is set to read-only.

This patch allows setting individual filesets as read-only. Depending
on which fileset applies during a client mount, the fileset's
read-only flag is used to force the mount as read-only. This is done
by setting an -EROFS error code server-side without setting the root
FID if the MDT connection does not match the fileset's ro-flag.

Note, the global nodemap "readonly_mount" property overwrites the
individual fileset read-only flag.

This patch adds the optional "--ro" flag to "lctl
nodemap_fileset_add". "lctl get_param nodemap.NM_NAME.fileset" was
modified to visualize the flag per fileset:

nodemap.c0.fileset=
[
 { primary:     /remote, mode: ro },
 { alternate:   /home },
 { alternate:   /data, mode: ro }
]

Sanity-sec "27e" was added to test this flag. Sanity-sec "27d" was
extended to explicitly test for a read-write mount.

Important implementation details:
The root fid retrieval is now done earlier in
"client_common_fill_super()". Previously, "md_get_root()" was called
after the client already connected to an MDT and OST. Since setting
the mount-point to read-only requires reconnecting to the MDT with the
corresponding read-only flag, "md_get_root()" is now called after
connecting to an MDT (before connecting to an OST) and after the
nodemap may have been set to read-only due to the "readonly_mount"
property.

Further, checking whether subdirectory mounts are server-side
available with the "OBD_CONNECT_SUBTREE" connect flag is now done
after the "MDS_GET_ROOT" RPC is sent (instead of before as
previously).  This guarantees that the connect flags are set and
"OBD_CONNECT_SUBTREE" can be checked.

Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I25e97e4c8da6318ec6d96a7797392257a4947e6b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57829
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>