LU-18687 compat: move hash/rhash to lustre_compat
Migrate the backported hash and rhashtable code to
lustre_compat.

The backported linux-hash.c code is only needed for
CentOS 7.5, so drop it.

Eventually, all of the Lustre/LNet compatability code
will live in lustre_compat - maintaining a clear
separation from the functional code in Lustre and LNet.

Test-Parameters: trivial
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: I73add2b5c2ad8ccc46c252f1066c7f20dc12cc39
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58157
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>