LU-11850 obd: support the rest of "stats" with Netlink
Migrate the remaining "stats" files to the debugfs Netlink API
except for the exports stats. Its is possible that we lack
a kobject and a debugfs dentry so we can end up in a case
that we can't derive a name. So change the API to supply
the stat source name instead. Update the stats packet size
calculate based on the new debugging info in the function
lnet_genl_parse_list().

Test-Parameters: trivial
Change-Id: If52dfb2807cbdcd9a24e9334edfa2101a8483fdd
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57305
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>