LU-18760 dkms: race on clobber and create of modules.order
DKMS builds fail occasionally with an error:

cat: /var/lib/dkms/.../build//modules.order: No such file or directory
  MODPOST /var/lib/dkms/.../build/Module.symvers

This appears to be a make bug trying where a path with //
is not understood correctly.

Remove the unnecessary injection of / in the list of SUBDIRS
to be built in when ldiskfs is not enabled.

Test-Parameters: trivial
HPE-bug-id: LUS-12672
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I6dda02133115076b076e6adf2ebabd10895af643
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58261
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Caleb Carlson <caleb.carlson@hpe.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>