LU-18770 lfs: mirror extend '-N=' get correct mirror number
It does not count the existing mirror number correctly in mirror
extend with "-N=<total_mirror_count>" option, since the raw layout
format stores the actual (mirror_count - 1) in its mirror count field,
so that plain or non-mirrored PFL file which has 0 in the mirror count
field logically has 1 mirror in it.

This patch fixes this error.

Fixes: a796b10fa2 ("LU-16446 utils: specify total count for mirror extend")
Signed-off-by: Bobi Jam <bobijam@whamcloud.com>
Change-Id: Ic1385a6b6e9adc191380d56ce95e422aa951734f
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58290
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Frederick Dilger <fdilger@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>